added PlannerUI project

+++ BuildingPlanner.Core/Abstraction/Shape.cs
@@ -2,6 +2,7 @@ namespace BuildingPlanner.Core.Abstraction
 {
 	public interface Shape
 	{
-		double Square { get; }
+        
+        float Square { get; }
 	}
 }
